Fix bugs in last update

+++ shap/common.py
@@ -288,7 +288,7 @@ def approximate_interactions(index, shap_values, X, feature_names=None):
                     v += abs(np.corrcoef(shap_ref[j:j + inc], val_other[j:j + inc])[0, 1])
         val_v = v
 
-        val_other = np.isnan(X[inds, i][srt])
+        val_other = np.isnan(X[inds, i][srt].astype(np.float))
         v = 0.0
         if not (i == index or np.sum(np.abs(val_other)) < 1e-8):
             for j in range(0, len(x), inc):
